libs/testing/pyproject.toml
[project]
name = "neuraprop-testing"
version = "0.1.0"
description = "NeuraProp AI shared test utilities"
requires-python = ">=3.12"
dependencies = [
    "neuraprop-core",
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "factory-boy>=3.3.0",
    "faker>=33.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/neuraprop_testing"]

[tool.uv.sources]
neuraprop-core = { workspace = true }
